requirement.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-dotenv
jose
